requirements.txt
Jinja2
MarkupSafe
PySocks
allure-pytest
allure-python-commons
ansi2html
async-generator
atomicwrites
attrs
certifi
cffi
chardet
charset-normalizer
colorama
configparser
crayons
docutils
execnet
greenlet
h11
htmlmin
idna
importlib-metadata
iniconfig
more-itertools
msedge-selenium-tools
outcome
packaging
pip
playwright
pluggy
py
pycparser
pyee
pyparsing
pytest~=7.1.2
pytest-base-url
pytest-forked
pytest-playwright
pytest-reporter
pytest-reporter-html1
python-dotenv
python-slugify
requests
selenium~=4.7.2
setuptools
six
sniffio
sortedcontainers
text-unidecode
toml
tomli
tqdm
trio
trio-websocket
typing-extensions
urllib3
wcwidth
webdriver-manager
websockets
wheel
wsproto
xlrd
zipp
Pillow